@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace RVSHidraulica
 {
@@ -17,7 +18,6 @@ namespace RVSHidraulica
         public Startup(IConfiguration configuration)
         {
             Configuration = configuration;
-
             Helper.MailHelper.Configuration = configuration;
         }
 
@@ -34,11 +34,19 @@ namespace RVSHidraulica
             });
 
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-        }
+            services.AddMvc(option => option.EnableEndpointRouting = false)
+                .SetCompatibilityVersion(CompatibilityVersion.Latest);
+        //    services.AddWebOptimizer(pipeline =>
+        //        {
+        //            pipeline.AddCssBundle("/css/bundle.css", "css/*.css");
+        //            pipeline.AddJavaScriptBundle("/js/bundle.js", "js/*.css");
+        //        }
+        
+        //);
+    }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -47,11 +55,12 @@ namespace RVSHidraulica
             else
             {
                 app.UseExceptionHandler("/Home/Error");
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
+                //// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+                //app.UseHsts();
             }
 
             app.UseHttpsRedirection();
+            //app.UseWebOptimizer();
             app.UseStaticFiles();
             app.UseCookiePolicy();
 
