@@ -16,7 +16,7 @@
     </div>
     <div class="mt-5 carousel w-100">
         <div>
-            <img src="@Url.Content($"~/img/c1.jpg")">
+            <img src="@Url.Content($"~/img/c12.png")">
         </div>
         <div>
             <img src="@Url.Content($"~/img/c2.jpg")">
@@ -31,7 +31,7 @@
     <div class="deg d-flex align-items-center">
         <p class="text-right col col-md-6 ml-auto mr-5 font-weight-bold">
             Nos ponemos a vuestra
-            disposición para ofrecerles las más amplias gamas de soluciones a sus necesidades.
+            disposición para ofrecerles la más amplia gama de soluciones a sus necesidades.
         </p>
     </div>
 </div>
