@@ -4,12 +4,12 @@
     <div class="mx-5 mt-5">
         <h1 class="text-center font-weight-bold">Cilindros hidráulicos</h1>
         <div class="d-md-flex mt-4 justify-content-center">
-            <div class="card col mx-3 col-md-5 p-0">
+            <div class="card col mx-md-3 col-md-5 p-0 mb-md-0 mb-3">
                 <img src="@Url.Content("~/img/c1.jpg")" width="100%" class="mx-auto">
                 <div class="card-body">
                     <h5 class="card-title text-center font-weight-bold">Cilindros</h5>
                     <p class="card-text">
-
+                        Simple y doble efecto, de acuerdo al requerimiento del cliente.
                     </p>
                 </div>
             </div>
