@@ -4,28 +4,27 @@
 <link rel="stylesheet" href="@Url.Content("~/css/valvulas.css")"/>
 <div class="mx-5 mt-5">
     <h1 class="text-center font-weight-bold">Válvulas direccionales y de control</h1>
+    <h4 class="text-center my-3">
+        Válvulas modulares y en línea para todo tipo de industria.
+    </h4>
     <div class="d-md-flex mt-4 justify-content-center">
-        <div class="card col mx-3 col-md-5 p-0">
+        <div class="card col mx-md-3 col-md-5 p-0 mb-md-0 mb-3">
             <div class="d-flex justify-content-center modulares">
                 <img src="@Url.Content("~/img/foto4.jpg")" width="50%" class="mx-auto">
             </div>
-            <div class="card-body">
+            @*<div class="card-body">
                 <h5 class="card-title text-center font-weight-bold">Válvulas modulares</h5>
-                <p class="card-text">
-
-                </p>
-            </div>
+            </div>*@
+            <h5 class="card-title text-center font-weight-bold py-3">Válvulas modulares</h5>
         </div>
-        <div class="card col mx-3 col-md-5 p-0">
+        <div class="card col mx-md-3 col-md-5 p-0 hid-mov mb-md-0 mb-3">
             <div class="d-flex justify-content-center">
                 <img src="@Url.Content("~/img/foto5.jpg")" width="100%" class="mx-auto">
             </div>
-            <div class="card-body">
-                <h5 class="card-title text-center font-weight-bold">Válvulas hidráulicas movil</h5>
-                <p class="card-text">
-
-                </p>
-            </div>
+            @*<div class="card-body">
+            <h5 class="card-title text-center font-weight-bold">Válvulas hidráulicas movil</h5>
+        </div>*@
+            <h5 class="card-title text-center font-weight-bold py-3">Válvulas para hidráulica movil</h5>
         </div>
     </div>
 </div>
