@@ -7,8 +7,8 @@
 <div class="mx-5 mt-5">
     <h1 class="text-center font-weight-bold">Servicios</h1>
     <div class="d-md-flex  mt-4 justify-content-between">
-        <div class="card col col-md-5 p-0">
-            <img src="@Url.Content("~/img/reparaciones.jpg")" class="card-img-top" alt="">
+        <div class="card col col-md-5 p-0 mb-md-0 mb-3">
+            <img src="@Url.Content("~/img/bombas-hidraulicas.png")" class="card-img-top" alt="">
             <div class="card-body">
                 <h5 class="card-title font-weight-bold">Reparaciones</h5>
                 <p class="card-text">Reparamos bombas a pistones, paletas y engranaje nacionales e importadas, motores hidráulicos, válvulas y cilindros.
@@ -16,7 +16,7 @@
                 </p>
             </div>
         </div>
-        <div class="card col col-md-5 p-0">
+        <div class="card col col-md-5 p-0 mb-md-0 mb-3">
             <img src="@Url.Content("~/img/foto3.jpg")" class="card-img-top w-75 mx-auto" alt="">
             <div class="card-body">
                 <h5 class="card-title font-weight-bold">Centrales hidráulicas</h5>
