@@ -49,7 +49,7 @@
                             <div class="dropdown-divider"></div>
                             <a class="dropdown-item" href="@Url.Action("ManguerasYAccesorios","Productos")">Mangueras y accesorios</a>
                             <div class="dropdown-divider"></div>
-                            <a class="dropdown-item" href="#">Accesorios hidráulicos</a>
+                            <a class="dropdown-item" href="@Url.Action("Accesorios","Productos")">Accesorios hidráulicos</a>
                             <div class="dropdown-divider"></div>
                             <a class="dropdown-item" href="@Url.Action("Valvulas","Productos")">Válvulas direccionales y de control</a>
                         </div>
@@ -66,17 +66,17 @@
         @RenderBody()
     </main>
 </div>
-<footer class="d-flex justify-content-around align-items-center mt-5">
-        <div class="d-flex flex-column text-center">
+<footer class="d-flex flex-md-row flex-column justify-content-around align-items-center mt-5">
+        <div class="d-flex mb-md-0 mb-3 flex-column text-center">
             <div class="text-white"><i class="fas fa-phone mr-2 text-white"></i> Teléfono</div>
             <span class="text-white">15-6369-5476</span>
             <span class="text-white">2137-1561</span>
         </div>
-        <div class="d-flex flex-column text-center">
+        <div class="d-flex mb-md-0 mb-3 flex-column text-center">
             <div class="text-white"><i class="far fa-envelope mr-2 text-white"></i> Email</div>
             <div><a href="mailto:rolando@rvshidraulica.com" class="text-white">rolando@rvshidraulica.com</a></div>
         </div>
-        <div class="d-flex flex-column text-center">
+        <div class="d-flex mb-md-0 mb-3 flex-column text-center">
             <div class="text-white"><i class="fas fa-map-marker-alt mr-2 text-white"></i> Dirección</div>
             <div class="text-white">Concejal Nicolás Defilippi 1253, Ituzaingó, Buenos Aires</div>
         </div>
